requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --output-file requirements.txt pyproject.toml
ansicon==1.89.0
    # via jinxed
beautifulsoup4==4.14.2
    # via izneo-get (pyproject.toml)
blessed==1.20.0
    # via inquirer
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via
    #   pytest
    #   tqdm
editor==1.6.6
    # via inquirer
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
inquirer==3.4.1
    # via izneo-get (pyproject.toml)
jinxed==1.3.0
    # via blessed
numpy==2.2.6
    # via
    #   izneo-get (pyproject.toml)
    #   opencv-python
opencv-python==4.12.0.88
    # via izneo-get (pyproject.toml)
packaging==24.1
    # via pytest
pillow==10.4.0
    # via izneo-get (pyproject.toml)
pluggy==1.5.0
    # via pytest
pycryptodome==3.23.0
    # via izneo-get (pyproject.toml)
## pytest has been moved to optional/dev in pyproject.toml
readchar==4.2.0
    # via inquirer
requests==2.32.5
    # via izneo-get (pyproject.toml)
runs==1.2.2
    # via editor
six==1.16.0
    # via blessed
soupsieve==2.5
    # via beautifulsoup4
tqdm==4.67.1
    # via izneo-get (pyproject.toml)
typing-extensions==4.15.0
    # via beautifulsoup4
urllib3==2.2.2
    # via requests
wcwidth==0.2.13
    # via blessed
xmod==1.8.1
    # via
    #   editor
    #   runs
